last submission

+++ src/app/employee-list/employee-list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, OnInit, Renderer2 } from '@angular/core';
 import { FormBuilder, FormGroup, FormControl } from '@angular/forms'; // Import FormControl
 import { EmployeeService } from '../service/employee.service';
 import { EmployeeFormComponent } from '../employee-form/employee-form.component';
@@ -17,7 +17,8 @@ export class EmployeeListComponent implements OnInit {
   constructor(
     private employeeService: EmployeeService,
     private formBuilder: FormBuilder,
-    private employeeform: EmployeeFormComponent
+    private employeeform: EmployeeFormComponent,
+    private el: ElementRef, private renderer: Renderer2
   ) {
     this.form = this.formBuilder.group({
       
@@ -39,10 +40,16 @@ export class EmployeeListComponent implements OnInit {
   loadData() {
     this.employeeService?.loadEmployeeData()?.subscribe((data: any) => {
       this.employeeData = data;
+      console.log(this.employeeData.length)
       
     });
   }
-  updateButton(id: number){
+  updateButton(elementId: String,id: number){
+    const element = this.el.nativeElement.querySelector(`#${elementId}`);
+    if (element) {
+      const yOffset = element.getBoundingClientRect().top + window.pageYOffset;
+      window.scrollTo({ top: yOffset, behavior: 'smooth' });
+    }
     this.employeeService.searchEmployeeData(id).subscribe((data: any) => {
        this.updateForm = true     
        this.form.setValue({
@@ -57,13 +64,13 @@ export class EmployeeListComponent implements OnInit {
     });
   }
   updateEmployee(){
-    console.log('are we ghe'+this.form);
-    
     const formData = this.form.value;
-    this.employeeService.updateEmployeeData(formData).subscribe(
+    this.employeeService.updateEmployeeData(formData)?.subscribe(
       (response) => {
-        console.log('Response:', response);
+        this.updateForm = false
         this.form.reset();
+        this.loadData();
+        alert('Employee Record Updated Successfully');
       },
       (error) => {
         console.error('Error:', error);
@@ -72,13 +79,24 @@ export class EmployeeListComponent implements OnInit {
   }
   deleteEmployee(id: number){
     this.employeeService.deleteEmployeeData(id).subscribe((data:any)=>{
+      alert('Employee Record Deleted Successfully');
       this.loadData()
     })
   }
 
   deleteAllEmployee(){
+    if(this.employeeData.length>0){
     this.employeeService.deleteAllEmployeeData().subscribe((data:any)=>{
+      alert('All Employee Records Deleted Successfully');
       this.loadData()
-    })
+    })}
+    else{
+      alert('No Employee Records to Deleted');
+    }
+  }
+
+
+  cancelButton(){
+    this.updateForm = false;
   }
 }
